Move Voting and Leaderboard components into pages folder
